pyproject.toml
[tool.poetry]
name = "rayim"
version = "1.8.3"
description = "Fast image compression for large number of images with Ray library"
authors = ["Mohammad Alyetama <malyetama@pm.me>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
tqdm = "^4.6"
Pillow = "^8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
rayim = 'rayim:rayim.main'
